+++ db/src/main/java/com/pyteam/db/mbg/mapper/Ad04Mapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface Ad04Mapper {
-    int countByExample(Ad04Example example);
+    long countByExample(Ad04Example example);
 
     int deleteByExample(Ad04Example example);
