+++ db/src/main/java/com/pyteam/db/mbg/mapper/Ad06Mapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface Ad06Mapper {
-    int countByExample(Ad06Example example);
+    long countByExample(Ad06Example example);
 
     int deleteByExample(Ad06Example example);
